update asset list, improve build error reporting, fix math warning

+++ Code/Engine/Graphics/Graphics.d3d.cpp
@@ -215,7 +215,7 @@ void eae6320::Graphics::SetEffect( Effect & effect, const Matrix4 local2world )
 	D3DVIEWPORT9 viewport;
 	result = effect.parent->GetViewport(&viewport);
 	assert(SUCCEEDED(result));
-	float fov = std::atan(1) * 4 / 3;
+	float fov = std::atanf(1) * 4 / 3;
 	float aspect = static_cast<float>(viewport.Width) / static_cast<float>(viewport.Height);
 	Matrix4 screenmat = ScreenTransform(fov, aspect, 0.1f, 100.0f);
 	const D3DXMATRIX * mat3 = reinterpret_cast<const D3DXMATRIX *>(&screenmat);
